[~TASK] FLOW3: Doubled speed of proxy building in Development context
Improved speed of getInterfaceNamesImplementedByClass() in the Reflection
Service by using PHP's native function. Previously array_search was
called more than 4,000,000 times in this method.

This is not a real solution, but at least a quick relief. We need to
take more measures and do more profiling as outlined in the related
issue.

Relates to: #9568

Change-Id: If6bce5c386b6b8ce1c51ec4a49386c69bdfc1560