[+BUGFIX] FLOW3 (Persistence): Safeguard for PDO credentials and DSN
The username, password and DSN for the PDO backend could, until now,
be NULL without causing any trouble. More correct though is an empty
string if they shouldn't be specified. In addition the old approach
caused an exception since recently.

Now the backend checks for the type of the credentials and the default
setting was adjusted to empty strings instead of NULL.

Change-Id: Ic551bf5985502135b48414d77c4611859fbef77d
Resolves: #9679