[+BUGFIX] FLOW3 (Persistence): Make property thawing order independent
Thawing of objects could fail if the same object was referenced inside
two properties of an object but the order in which these were returned
from the persistence backend differed from the order in the classschema.

The last fix introduced in Change I5bd71c9a contained a bug: properties
of type object which originally contained NULL, contained an empty array
after thawing. This patch should fix that problem.

Change-Id: I6b4635081f00d90b95ecef0555265803ce7c92c8
Fixes: #9684