[+BUGFIX] FLOW3 (Persistence): Repository functions pass fetch mode to execute() explicitly
Two of the repository tests failed with a warning because it uses the
QueryInterface for creating mocks and in this interface execute() is
declared with a non-optional parameter $fetchMode. In the Query implementation
however the parameter is optional and therefore the methods in Repository
could call execute() without an argument if Query was real.

This patch makes all execute() calls in Repository passing the fetch mode
explicitly.

Change-Id: I20b214eb792031ba1a7068925e40318e1bdfc6d8