[~TASK] FLOW3 (MVC): Enhanced capabilities and configuration for JSON view
The JSON view was configurable through a special YAML file. However, this
was not practical in many situations. Now the configuration for renderable
properties of the JSON view's variables is done programatically.

The view also had difficulties to render certain structures containing
objects. For example getter methods were called even though their property
names are in the list of excluded fields. This could cause problems with
getters which require arguments, such as getNode($path). This is solved
by this patch.

Change-Id: I7788faa16b88636510bd9a90473eebb84cd9cd68