[!!!][~TASK] FLOW3 (Persistence): Refactor QueryResultProxy into QueryResult
Query->execute() now only returns QueryResultInterface instances, those
results have a toArray() method to get the result as array of objects
and a getFirst() which returns the first object found.

The count() method has been removed from the Query class and is now
expected to be used on the query result, so that each $query->count()
now should read $query->execute()->count().

Change-Id: Ib3fb13bce47772b34992cb357b2b4493850b832e
Resolves: #9710