[+BUGFIX] FLOW3 (MVC): Request::setControllerObjectName now works with multiple subpackages.
Before this change, calling "setControllerObjectName" did not detect the subpackage correctly,
if the class name was like F3\Package\Subpackage1\Subpackage2\Controller\MyController.
Now, it correctly detects "Subpackage1\Subpackage2" as the subpackage.

Additionally, the order inside the regular expression was changed,
which should slightly increase the performance in case there is no
Subpackage in the Controller Object Name.

This change is needed for the Widget API to work.

Change-Id: Ie753385f08c43680f15a1acdeffcf6f0c070bb5d