[+BUGFIX] FLOW3 (Validation): Fix building of base validators
The conjunction validators built for base validation of models did not
behave correctly, validation of contained objects was not added at all,
if no @validate annotation was found (#9862).

In addition validators for nested objects (A contains B which contains
A) could not be built due to ending up in endless recursion (#9855).

Change-Id: Id62828271312adb5c8e21c000a061b008b4596ec
Fixes: #9855
Fixes: #9862