[+BUGFIX] FLOW3 (Property): Allow Object Converters to use any input format
Theoretically Object Converters should be able to support either arrays,
string or numbers as input format (if they wish). However, due to the program
flow in transformToObject() an Object Converter would only be invoked if the
source was an array.

This patch fixes this and also gives matching object converters precedence
before the Property Mapper's internal mapping functions (such as identity
array to object).

Change-Id: I16fc419ca40e28ae2605e06bde8dcbdd166b5231