[+FEATURE] FLOW3 (Persistence): Add default ordering for repositories
When using findAll() or findBy*() the order of the result cannot be
customized without writing your own methods to inject the ordering.

And if a repository wanted to apply some default ordering to it's own
methods, it still had to either use setOrderings() all over the place or
override createQuery() to centralize that.

Now you can use setDefaultOrderings() on any repository or define
$defaultOrderings in a repository directly. The setting will be picked
up by createQuery() right away.

Change-Id: I261aec2a5e024028bbbd0efad51a5953559f6c1f
Resolves: #9709