[~TASK] FLOW3 (Persistence): Add metadata to proxy from object data
For other storage backends (like CouchDB), it is interesting to store
additional metadata for objects inside the DataMapper. This could be
used to store a document revision inside an entity for updates.

If the metadata is not set by the backend when constructing the object
data, the change will have no effect.

Change-Id: I30e13c9760b1d893e34e006a2091c36b660cd942