[~BUGFIX] FLOW3 (Persistence): Fix multivalue dirty checking
The dirty checking of an array with objects always returned TRUE,
because an if statement contained a check for TRUE at the wrong
position. The unit test did not cover this case, because the parameters
of the mock method invocation were not checked.

Additionally added required parameter for ArrayObject in two tests.

Change-Id: I1d84144e9de4ee3e02542992d87bcc997b599535