[+BUGFIX] FLOW3 (Object): Register DateTime at the object management
The PHP-built-in DateTime class is unknown to the Object Manager, yet
it is used at several places, such as the persistence framework. Any
try to get further information about this class (such as what scope
it has), results in an exception thrown by the Object Manager.

This patch pre-registers DateTime and therefore adds Object Manager
support for it.

Resolves: #10341
Change-Id: I2fa63525fd40b838d8bc1464b6b86357e057103e