[+FEATURE] FLOW3: Functional Tests Support
This changeset adds support for functional tests in related
parts of FLOW3, such as the Class Loader, Object Management and
Package Management.

The new context "Testing" is now supported by FLOW3.

Class files which are located below Tests/Functional/ can now
be automatically registered at the object manager by enabling the
registerFunctionalTestClasses option (Settings.yaml). This is now
the default behavior in the Testing context.

The bootstrap now provides a getObjectManager() method for the
functional base testcase to fetch the object manager instance.

Resolves: #10425

Change-Id: Ia13a8acb63d67bd17c525dacbe5253ceb58c4f1e