[+BUGFIX] FLOW3 (Cache): Work around PHP __halt_compiler() namespaces bug
The PhpFrontend added a __halt_compiler() instruction to the end of each
cache file to ignore following meta data needed by the cache framework.
This resulted in wrong namespace scopes for certain - originally global -
classes.

Instead of __halt_compiler() we now use a simple "#" comment which should
work equally fine with the current kind of metadata we are using.

Change-Id: I698c15e1053ae87337bb3c914c9c98de35a2e536
Resolves: #10428