[+FEATURE] FLOW3 (MVC): Support __call() method for actions
The ActionController checks if the resolved action method actually
exists before it is called. In the TYPO3 package however we have
an ErrorController which tries to catch any original action call
and forwards it to the indexAction. This is achieved by a magic
__call() method.

This patch now checks if an action method is callable and not if
it really exists.

Change-Id: Ib9407b25767501e225b42811211c12a99395188a