[~TASK] FLOW3 (Cache): Rename Redis Cache Backend
This renames the cache backend formerly named "PhpredisRedisBackend"
to "RedisBackend".

If there should be, at some point in the future, a second implementation
of a Redis backend, or another way for accessing Redis should be supported,
then this should be solved with a configuration option for this RedisBackend.

Relates to: #9017

Change-Id: I5cc2988b424f490db8eddad8023af7b2fb4fe9ca