[+BUGFIX] FLOW3 (Package): Remove is_dir check for packages directory
Upon a fresh git clone, there is no Application directory as git does
not track empty folders. This caused an exception to be thrown. Removing
the is_dir() check solves the problem, and the directory is created
anyways if it does not exist. Furthermore, the exception was misleading
because the user could have specified another path than
Packages/Application.

Change-Id: I7cbd0aae9b168de26710b406440d178635761e90