[~TASK] FLOW3 (Security): Support additional tokens
Other tokens than the UsernamePassword token could need
a request instance when updating the credentials. Another problem
is, that the LoggingAspect is tied to the UsernamePassword token
and always expects credentials with a username array key. The token
interface now expects a __toString method to return a string
representation of a token for logging purposes.

Change-Id: I8e0d70a7a27a1741b7df3b5fccfd4672abae4f82