[~TASK][!!!] FLOW3 (Reflection): Improved speed of reflection service
This patch accelerates the initialization of the ReflectionService
in production context. Instead of checking for emerged classes or
generally for available class names, the service will now reflect
classes only once (when the reflection cache is empty) and won't
try it again (if detectClassChanges is disabled, such in production
context).

Another measure to speed up initialization was to not cache the
method DocComments anymore. Saves more than 1MB of caching data.

NOTE: The initialize() method of the ReflectionService now has
a different signature. If you mocked the ReflectionService in one
of your tests, these tests probably need to be adjusted.

Change-Id: I75cce21c7a9d6b58f41d92ecb0948245a43f812d