[!!!][+FEATURE] FLOW3 (Security): Allow authentication of multiple tokens
Currently the authentication provider manager allows only the
authentication of one token (the first that authenticates) or all active
tokens. For a flexible multi factor authentication it is suitable to
have an additional strategy and authenticate at least one but as many
tokens as possible.

Note: If you changed the setting for authenticateAllTokens to TRUE you
must now set authenticationStrategy: allTokens instead!

Change-Id: I1837764525841d8a9100c4635a4f3294a5730ab6