[~TASK][!!!] FLOW3: Moved and renamed base test cases
This moves the base test cases for unit, functional and selenium tests
from the Testing package to the FLOW3 package. The new unit test case
is now known as F3\FLOW3\Tests\UnitTestCase. The functional and selenium
test cases are called FunctionalTestCase and SeleniumTestCase in the
same namespace.

For backward compatibility, a legacy unit base test case using the old
class name and namespace is available. Test cases of third-party packages
should, however, switch to the new class names as soon as possible.

This also removes the PdoInterface from the global location to the GenericPdo
test directory.

Change-Id: I75ada37fc887dcdf8332ad3a5caa40e88260ab41