[+BUGFIX] FLOW3 (Security): Workaround for session concurrency problem
The FLOW3 session management seems to have concurrency issues.
This results in token objects being duplicated and an ever growing
session which results in an unusable system. This behaviour is
reproducible both in Development and Production context by firing
concurrent requests while being logged in.

The change works around the problem by making sure tokens of the same
class are never duplicated by indexing the token arrays in the
security context with the tokens' authentication provider name.

Change-Id: I2d4141e5730836da43f0c805878f2eec3e302a93