[+TASK] FLOW3 (Scripts): Add ResourcePointer migration script
The introduction of protected resources also brought ResourcePointer
objects and made Resource objects entities. This breaks existing data,
which can be migrated with this script.

If you are not using the default SQLite database, you need to edit the
connection parameter in the script.

BACKUP your data before using this!

Change-Id: I896184fa2a0efa329a4b4c9bffe4f8f93e53665e