[+BUGFIX] FLOW3 (MVC): Fix redirect for sub requests
When using redirect() in a controller processing a sub request
the additional arguments of the sub request were carried along.
This resulted in unexpected nested rendering (in TYPO3 the login
screen would be rendered into the section where the plugin triggering
the redirect was located).

redirect() will now determine the top level request and use that
to execute a proper redirect.

This patch also contains a fix which allows to use exernal URIs
with redirectToUri(). Previously all URIs passed to that function
were prepended by the base URI.

Change-Id: Ia410ac886ccf4a8c1f380ae25d691999586f2876