[~TASK] Clean up and enhance cache backends
If the frontend attached to the FileBackend is a PHP frontend, a ".php"
file extension will now be used for each cache entry. Most IDEs or
debuggers would not recognize the cached code otherwise.

Also removed the logging functionality from all backends because logging
is not available at the point when caches are used in the Bootstrap.

Moved Environment to the AbstractBackend.

Change-Id: I9783e629bc0148cade69acc4f819f7071d1b607e