[~TASK] Use common temp dir for each context
Previously FLOW3 used a dedicated directory for each SAPI type and some
more. Now that the Proxy Class Compiler must be able to generate code
which can be used throughout a context, this separation is not possible
(and not neccessary) anymore.

Instead each context has one temporary directory in which no further
separation is done.

Change-Id: I18c920b8eaaedd285a38a2289fd5052e1e61c225