[+FEATURE] New proxy class compiling mechanism
This commit features a new proxy class compiler which is able to produce
generic proxy classes. It also contains to uses of this mechanism: the
Depndency Injection Proxy Class Builder and the AOP Proxy Class Builder.

Proxy classes are built during a special compile step which is
implemented in the Bootstrap.

The Object Manager has been refactored too. Calling create() is now
deprecated and PHP's "new" operator should be used instead for
instantiating prototypes. As for the dependency injection functionality
some restrictions apply which will be outlined in the FLOW3 manual
later.

Change-Id: I0031253f452329c11d487002602e873d48cc792b