[!!!][~TASK] Make "prototype" default scope and additional scope checks
Prototype is now the default scope if nothing else has been defined in
the class (by a @scope annotation) or in the Objects.yaml. You'll need
to check your classes if they are still in the right scope.

This also adds another new behavior: if an object is specified by an
interface name, the object management will now check if there's a
@scope annotation in the implementing class. Also an error will be issued
if a @scope annotation is used in an interface.

Change-Id: I760237d9230b81fb55fc2db23ab16ee0871bf371