[~TASK] Improve compile support in Bootstrap
This is work in progress, but still should be merged into master if it
doesn't cause any problems. It stabilizes the CLI support for automatic
compilation and at least starts to cleanup the corresponding parts in
the Bootstrap.

This changeset also removes the Lock Manager.

While we might need a concept / mechanism similar to the Lock Manager,
the current implementation does not fit to the new compiling mechanism.

Change-Id: I2f0a7ba4cd9854828482ae08b8f189911c4e64bb