[+BUGFIX] Bugfix for "Replace readlink() by realpath()" commit
This is a bugfix for I2cc4485bf11741d1785c9c19381d622f03b4c0ff

I have the following setup:
 FLOW3Base/ -> the flow3 base distribution
 BlogExample/ -> BlogExample distribution
 BlogExample/Packages/Framework is a *symlink* to FLOW3Base/Packages/Framework

In this setup, realpath(BlogExample/Web/_Resources/FLOW3) resolves to
FLOW3Base/Packages/Framework/FLOW3/Resources/Public, and *not* to
BlogExample/Packages/Framework/FLOW3/Resources/Public, which is the
sourcePath.

The solution is to add realpath() on the source side as well.

Change-Id: Ie5445520d4752a4096cb20b5abbb13433ffe9121