[!!!][+FEATURE] Command line commands and optimized bootstrap
This introduces the new CommandController which can be used for providing
commands for the command line. The regular AbstractController and thus
the ActionController is not and should not used for CLI requests
anymore.

The bootstrap now handles command line and web requests more efficiently.

This commit also features a few core commands such as flush all caches
or compile.

Change-Id: I9edf26c266ca977c3d31ddb11453af04ca5b9616