[+BUGFIX] Reflection Service does not save emerged class reflections
Because the Reflection Service was not correctly connected to the
bootstrap shutdown emerged classes (which were not automatically
registered during compile time) were not saved into the cache and
therefore reflected again on every run.

This patch introduces new signals into the bootstrap and Core Command
controller and makes connects the Reflection Service's saveToCache()
method to one of the new signals.

This bug fix was auctioned at the T3BOARD11 bug auction by wildside.dk

Change-Id: I25d94f6ea84dfa84ea29aed830c71c3bbe183079
Resolves: #25607