[+BUGFIX] Fix compile command on Windows
On Windows there seems to be no reliable way to determine
the location of the PHP binary.
By adding the setting "core.phpBinaryPathAndFilename" the path
can now be overridden manually. It defaults to "%PHP_BINDIR%/php" which
should work on unix based systems.

Besides the format of setting environment variables is different on
Windows.

Change-Id: I480fb4e78f7610be88420bc928b636ef0d0c910a