[BUGFIX] Make links/redirects work in SubRequests
Web\SubRequest::getBaseUri() is never set and it makes no sense to have
a custom base URI in a sub request anyways.

But that leads to the situation that plugin redirects/links are always
relative which obviously doesn't work if you are on a sub page.

This is fixed by bubbling all URI getters and setters up to the parent
request

Change-Id: I2d3e14838f68b83e58fcdcf483226c8c422574e7