[+BUGFIX] Directory Iteration should skip hidden directories
The problem manifests if one uses SVN, which adds .svn
directories into places which are monitored by file monitoring.

Then, also files inside the .svn directories are monitored;
which crashes the cache system later, with an exception:
1233057359: "%CLASS%F3_MyPackage_.svn_text-base_Package" 
is not a valid tag for a cache entry.

Bugfix: Just skip over directories if they start with ".",
which is consistent as we do this with files as well.

Change-Id: I0d4c1ed601a7d55fc36c136b3f8fa3c246d0de7d