[TASK] Run Doctrine setup as CLI sub request
When FLOW3 needs to set up Doctrine, the proxy classes for entities need
to be used, because without them no AOP-introduced identifier properties
are available.

Thus we call the setup in a second sub request after FLOW3 has finished
the proxy compilation. In production context this is not done, but must
be manually executed from the command line:
 ./flow3 flow3:doctrine:update
 ./flow3 flow3:doctrine:compileproxies

Change-Id: I3f278762f553a1e2f7738fff1a86dfa09c4aa5ae