[+FEATURE] Basic help support for command line commands
This adds some basic help which displays a list of available commands
and a short description for each. This implementation is not optimized
for speed (collected information should be cached later on) and doesn't
provide detailed help for a single command, but it helps a bit already.

Also changed the descriptions of a few commands and renamed a signal
emitted by the CoreCommandController because the signal name ended
with "Command" and thus was listed as an available command in the
help index.

Change-Id: I9731925f2958e819c9c32a76bbca0174d1906cee