[+BUGFIX] Always lowercase file extensions in Resource
When uploading a resource with upper case characters in the
file extension (for instance "SomeFile.JPG") the saved
persistent resource will be saved with a lower cased file
extension ("<hash>.jpg"). This is correct, because it avoids
clashes for two files with the same content (= same hash).

This change makes sure that Resource::getFileName() also
contains the lowercased file extension.

Fixes #26972
Related #26965

Change-Id: I7b2a850e0302a03095b4e9ecb4e83bda6502b5c9