[TASK] Make getResourcePointerForHash() public
As in some cases ResourcePointer objects are created outside the
ResourceManager, they might conflict with ResourcePointer objects already
stored. As long as Doctrine doesn't support value objects and this method is
needed it should be public.

Although it's public, it's no part of the public API and disappears as soon as
value object handling is implemented.

Change-Id: I6f2e7225073a23a6960652fb7e97ec805cee558c