[TASK] Run Doctrine schema and proxy setup only if needed
With this change doctrine schema updates and proxy compilation are only
run if class files in .../Domain/Model/ have changed. This speeds up the
whole bootstrap process significantly for situations where the model
code is unchanged.

In Production context the update is run only if the caches were flushed,
this is consistent with FLOW3 proxy compilation behavior.

Change-Id: I846fc4c7eff4523947bb17f3963e518390547199