[BUGFIX] AOP: Don't call parent::__wakeup() twice
When defining a custom __wakeup() magic method in a class, it was called
twice in the proxy method. This fixes this by checking if the parent
defines a __wakeup method and only add the call code if this is not the
case. This also moves the parent call to be postParentCall instead of
preParentCall so the parent method is called after dependency injection.

This fixes behavior introduced by Change-Id I662593f77bc52bad7e9b708fb5a6d69f470aa15a

Change-Id: I3f5e98cc47619ab76362b53c9b225e6d4acbe65f