[BUGFIX] AOP proxies pass NULL arguments to parent
Because isset($array[$key]) returns FALSE if $array[$key] === NULL,
method arguments having NULL as value where not passed to the original
method and the position of the remaining arguments was shifted to the
left resulting in unexpected behavior. This fixes this by using
array_key_exists() instead of isset().

Change-Id: I09ac6b0d85c98faca18c95253bea3be272d3a5b8