[BUGFIX] RouteParts must always return strings
With resolution of #26669 DynamicRouteParts have been extended in order
to support multidimensional arguments.
While this makes sense for match, it leads to an exception if the
routePart value is no string after $routPart->resolve() has been
called because the value can't be turned into a corresponding URI
portion then.

Change-Id: If2a4b6d937d03acb03cf1c25846febe223183981
Related: #26669