[BUGFIX] Fix runtime evaluations with global objects
After the vendor namespace changes, the configuration manager
didn't work as supposed and global runtime evaluations on aspects
or policies failed. This change switches to the now available
getSettingsByPath method of the ObjectManager which should also go
to the interface. Also some cleanups are applied and a functional
test for global objects is added.

Change-Id: Id3b7c348b1684f4774419203e68ed1e71c7a1be9