[FEATURE] Warm up caches
This adds a new command "flow3:cache:warmup" which warms up caches
in order to prepare the application for the first requests when going
live.

The cache:warmup command emits a signal which allows other application
parts and packages to execute additional tasks as preparation for the
big rush.

This change set also contains a small improvement for compile time
command controller registration: it is now possible to register either
a whole controller (by using a wildcard) or single commands as compile
time.

Change-Id: I04d2812ba97877e3bc11c1b2cc1bc7eedc823a80