[BUGFIX] It should be possible to disable Doctrine integration
This change introduces a setting to control the integration of Doctrine
in the bootstrap of FLOW3. For non-Doctrine persistence backends no
Doctrine code should be invoked, otherwise warnings about missing
driver or driverClass options will occur. And there is no need to
do proxy code compilation for Doctrine if it's not used.

Change-Id: I0ae80d2106c758d12242805ed20becb696e75da8
Resolves: #28228