[BUGIFX] Fix table name inference for packages without vendor name
The current way of guessing the table name for Doctrine metadata
was fixed to a specific package naming. Multiple vendor names
("Com.MyCompany.MyPackage") and a missing vendor name were
not supported correctly. This change fixes these problems and
is compatible to the current algorithm for package keys like
"TYPO3.FLOW3" or "TYPO3.Party".

Change-Id: I276149c5d651ba11d6179b56684a6a480efdef55