[FEATURE] Intelligently recognize boolean options
This adds some more intelligence into the command-line request building.
If a boolean option is specified without a value, a following unnamed
argument will be considered as such and not as the value of the boolean
option:

   kickstart:actioncontroller --force Acme.MyPackage

Because --force is boolean, Acme.MyPackage won't be considered as the
value of the force option. If a "boolean" value is specified though,
it _will_ be recognized as such:

   kickstart:actioncontroller --force yes Acme.MyPackage

Possible "boolean" values are:

   yes, no, y, n, 1, 0, true, false, TRUE, FALSE

Change-Id: I87ced3ccfafc1b5daf53a74397f7e80cb1e872f6