[!!!][FEATURE] Speed up ObjectAccess
This change improves performance of ObjectAccess::getProperty(),
ObjectAccess::getPropertyPath() and ObjectAccess::isPropertyGettable()
by roughly 100% by changing the order of type handling and reducing the
amount of type checks in internal calls.

Also this removes support for Closure evaluation, as it is not used anymore
within the core.
However this should not change the behavior in most cases, unit tests
are still running through.

Change-Id: Ia2e7bb2ec4c9dd9ca150572e25478685be90d203