[BUGFIX] Proxy code of advised abstract classes don't produce errors
The AOP ProxyClassBuilder produced invalid code when signals were
defined in an abstract class. This was due to invalid proxy class
code and a wrong initialization procedure.

This change fixes the issue by making sure that the advice information
is also built for parent proxy classes when sub classes are used.

Additionally this change set contains functional tests for the
AOP proxy and the signal slot mechanism.

Change-Id: I6480321c117dc0eb264fda45a952d27505156f82
Fixes: #28638