[BUGFIX] Set proper file permissions with setfacl
Previously the ACLs set by the setfilepermissions command
did not really grant access to the web server's user.

This could be reproduced by changing the owner of all directories
to root:root, then running the script and trying to run FLOW3
from the web which then fails (cannot write to _Resources etc).

Change-Id: I76d91010508713568c99af1a0ba6b047e91d425b