[TASK] Require Windows 6.0 or higher
The use of symlinks in FLOW3 is an essential feature. Without
it, certain parts of the caching and resource management would
be unacceptably slow.

Therefore the minimum requirement on Windows platforms is now
Windows Vista, Windows Server 2008 or higher.

This change set also makes sure that all code parts in the Bootstrap
which exit do send a correct exit code.

Change-Id: I4e75a8351174ad68bde83101373f6a5b9c059043
Resolves: #11683