[TASK] Throw exception if type of relation is not defined
If a relation to a non standard datatype is defined in a model you have to define the type
of the relation (OneToOne, OneToMany...). If not doctrine tries to convert the related
model to a string and throws an exception.

Add an extra exception for this special case which gives a hint what's wrong.

Change-Id: I0252e4ea07854908d5a04c3d083e560174f9ca77
Resolves: #28868