[BUGFIX] Remove unnecessary chmod for PackageStates
The Package Manager tried to adjust the permissions of
the PackageStates.yaml file via chmod.

This should not be necessary anymore and could even lead
to irritating error messages for some server setups.
In general, the file permissions should be set by the
setfilepermissions command in a way that even a newly
created PackageStates.yaml has the correct permissions
set.

Change-Id: I7d235228223331ccd6d7cf2f6981f26a0cd95719
Resolves: #28814