[FEATURE] Simple remote package import support
This adds a new command package:import which allows for importing
a package from git.typo3.org. Currently only packages of the TYPO3
vendor namespace are supported and it requires that the git command
line tool is installed and executable by PHP.

Note that it is very likely that this feature won't stay like this in
future versions of FLOW3.

Change-Id: I918471c7be1ad1fe2b387325b08911695591890a
Relates: #29489