[!!!][BUGFIX] DateTimeConverter returns Errors instead of throwing exceptions
As described in the Property Mapper reference manual, DateTimeConverter should
only throw exceptions when detecting *developer failures*. If *user failures*
happen, such as invalid input, an \TYPO3\FLOW3\Error\Error object should instead
be returned, as this is shown to the end-user in a friendly way, and the original
form is redisplayed.

Furthermore, DateTimeConverter must convert an empty date string ('') to NULL;
else it is not possible to define *optional* DateTime properties.

If you need a *required* DateTime property, use the @validate NotEmpty annotation
in your model.

Change-Id: Ie95602c44e9d86586353514cbde2cd06259f3b30
Resolves: #27806