[TASK] Improve message for compile errors
If the compile run fails, FLOW3 will display a long exception
message stating that either the PHP binary could not be executed
or something else went wrong.

This patch splits up the messages in two and checks if the binary
could be executed and only then gives a hint on configuring the
PHP binary path.

Change-Id: Ia9379496add1721a0b388a8665934a641caa470d