[BUGFIX] (Property): Enable mapping of Value Objects
If a value object is edited through a Fluid form, we should not try to fetch it
from persistence, but need to re-create it from scratch due do the value object
semantics.

Thus, we need to discard the __identity property in that case.

The change set also adds a functional test.

Resolves: #29905
Change-Id: Iae3eaafffab1b062ff8ef112cfa9aac9f931e46b