[BUGFIX] Invalid static resource symlinks are removed again
Whenever the resource publishing of FLOW3 finds an invalid
symbolic link in the static resources directory, it should
remove that link and re-publish the resources of the package
in question.

Because Utility\Files::removeDirectoryRecursively() could
not cope with symbolic links, this auto-cleanup failed.

This change set makes removeDirectoryRecursively() and
emptyDirectoryRecursively() symlink aware and sets its
behavior to not follow links but remove the links instead.

Note: change set I2d5c544e6db728594209294cfe779b8bad3c6d59
did not completely fix the issue. If the old symbolic link
pointed to nowhere, the publishStaticResources() method
failed.

Change-Id: I30a27d60a93de8aeb7d27ea8030df6748d4af4aa
Fixes: #29777