[FEATURE] Add generic Identity RoutePartHandler
This implements improved routing for persisted objects.

When an object type is specified in the routing configuration,
the new "IdentityRoutePart" is used to match/resolve the route
part.

By default the resolved URI representation (path segment) is
built from the identity properties of the domain model. If no
identity properties are defined, the technical identifier (UUID)
is used.

Additionally a custom URI pattern can be defined in the route.
For properties of type DateTime the date format can be specified
prefixed by a colon.

If a path segment already exists, the IdentityRoutePart handler
will append an increasing counter to ensure unique URIs.

Sample Route:
-
  uriPattern:    'posts/{post}'
  defaults:
    '@package':    'TYPO3.Blog'
    '@controller': 'Post'
    '@action':     'show'
    '@format':     'html'
  routeParts:
    post:
      objectType: TYPO3\Blog\Domain\Model\Post

Change-Id: I8fb5d4f4be3649e2f6307f83af2f3bbaa2191b5b
Resolves: #8736