[!!!][TASK] Create lower cased URIs by default
The routing framework is capable of creating case sensitive
URIs so it is possible to resolve different actions depending
on the case of URI segments.

This changes the behavior to convert URIs to lower case by
default as that is the most common requirement.

If you want to keep the case, you can achieve that by setting
toLowerCase to FALSE. This can be done for the whole route or
for specific Route Parts:

-
  name: 'Case sensitive Route'
  uriPattern:    'posts/{post}/{category}'
  defaults:
    '@package':    'TYPO3.Blog'
    '@controller': 'Post'
    '@action':     'index'
    '@format':     'html'
  toLowerCase: false

-
  name: 'Case sensitive Route Part'
  uriPattern:    'posts/{post}/{category}'
  defaults:
    '@package':    'TYPO3.Blog'
    '@controller': 'Post'
    '@action':     'index'
    '@format':     'html'
  routeParts:
    post:
      toLowerCase: false

Note: This is a breaking change in the sense that dynamic route
parts are now resolved to a lower cased string by default.

Change-Id: I572b51af2992e2599b37db3d0e5b31cbb395ece1
Resolves: #29977