[BUGFIX] CLI Help throws exception if related comments are not available
The flow3:help:help command parses @see annotations of the
commands and retrieves the method description of related
commands.
This leads to an exception if the target command is not
available (e.g. if the respective package is not loaded).

This change fixes this by replacing the short description
with "<i>Command not available</i>" in those cases.

Change-Id: I961cdeaf2a0928017bdfaec16bfcb014ee428076
Resolves: #30261