[BUGFIX] Routing does not always work if splitString is optional
If the route part following an identity route part optional
like "{post}(.html)", the identityRoutePart creates a correct
link but matching of the URI without the optional part fails,
if the dynamic part contains the split string (in this example
if the post contains the string ".html").
This change fixes this by always matching against the complete
routePath with an improved RegEx.

Resolves: #30929
Change-Id: Iae4ba2652c08e8a0ba468076ae5dec6008f0ea47