[BUGFIX] Static Route Parts are not reset
StaticRoutePart::resolve() sets the value to the route parts
name (= the corresponding URI part).
RoutePart::match() merges the route parts value into the
resulting request arguments. The value of the route part must be
reset to NULL before the route parth is processed.
The StaticRoutePart currently fails to do so. Usually this has
no effect because match() is always called before resolve() but
it fails when using FLOW3 with an app server like AiP.

Change-Id: Ide104658cb1c822e58768200092c30dd0b356e33
Resolves: #30542