[TASK] Support new annotation syntax
This adds support for the new annotation syntax in a way that aims
to change as little code as needed. It does this by adding support
for namespaced annotations to DocCommentParser, adding some methods
to read annotations using Doctrine Common to the ReflectionService
and adjusts some other parts of the code that needed to be changed
already.

Existing annotations in userland code can be migrated using the
migrate-annotations.php script. Run it without arguments to get
instructions.

Further changes to fully resolve #27627 will follow.

Change-Id: I4236995770943446410da517a9af9707d5d37e11
Related: #27627