[BUGFIX] make is_link() work around work for paths with trailing slash
On Windows \TYPO3\FLOW3\Utility\Files::is_link() compares the
given path with the resolved absolute path due to a bug in the
native is_link() implementation.
This fails if the given path ends on a slash because that is
stripped of by PHPs realpath() function.
This change strips of any trailing slash before comparing the
paths.

Change-Id: I94c53b5fe8dda82c05690cf02286091672d0650e
Resolves: #31061