[FEATURE] Persistence Manager knows about unmanaged, new objects
This patch enhances the persistence aspect and the related code in the
Persistence Manager so that getObjectByIdentifier() also returns those
objects which have been created but not yet added to any repository.

That allows, among other things, to render URIs pointing to not yet
persisted objects which have been carried around in a session container.

Resolves: #29578
Releases: 1.1
Change-Id: I5cb5f497b475600dc0faa448600bdd7af8927e7a